@@ -1,10 +1,10 @@
 source 'https://rubygems.org'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.0.2'
+gem 'rails', '4.1.2'
 
 # Use SCSS for stylesheets
-gem 'sass-rails', '~> 4.0.0'
+gem 'sass-rails', '~> 4.0.3'
 
 # Use Uglifier as compressor for JavaScript assets
 gem 'uglifier', '>= 1.3.0'
@@ -22,24 +22,34 @@ gem "cocoon"
 
 gem "paperclip", "~> 4.1"
 
-gem 'bootstrap-sass', '~> 2.3.1.3'
-gem 'bootswatch-rails'
+gem 'acts_as_commentable'
+
+gem 'wicked_pdf', github: 'mileszs/wicked_pdf'
+
+gem 'bootstrap-sass', '~> 3.2.0.1'
+
+gem "bootstrap-switch-rails"
+
+gem 'acts_as_list'
 
 # See https://github.com/sstephenson/execjs#readme for more supported runtimes
 # gem 'therubyracer', platforms: :ruby
 
+gem 'redactor-rails'
+
 gem 'paper_trail', '~> 3.0.2'
 
-gem 'cancan'
-gem 'cancan_strong_parameters'
+gem "pundit"
 
 # Use jquery as the JavaScript library
 gem 'jquery-rails'
+gem 'jquery-ui-rails'
 
 gem 'autogrow-textarea-rails', '~> 3.1.0'
 
 # Turbolinks makes following links in your web application faster. Read more: https://github.com/rails/turbolinks
 gem 'turbolinks'
+gem 'jquery-turbolinks'
 
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
 gem 'jbuilder', '~> 1.2'
@@ -65,6 +75,8 @@ group :production do
 end
 #gems for testing
 group :development, :test do
+  gem 'minitest'
+  gem 'database_cleaner'
   gem "rspec-rails"
   gem "factory_girl_rails"
   gem 'guard-rspec', require: false
