@@ -1,21 +1,42 @@
 class BooksController < ApplicationController
-  load_and_authorize_resource :book, :only => [:index, :show]
-  before_action :set_book, only: [:show, :edit, :update, :destroy]
-  before_filter :authenticate_user!
+  before_action :set_book, only: [:show, :edit, :update, :destroy, :publish, :unpublish, :complete, :uncomplete]
+  before_filter :authenticate_user!, :except => [:view, :download]
 
   # GET /books
   # GET /books.json
   def index
-    if can?(:manage, :all)
-      @books = Book.all
-    else
-      @books = current_user.books.all
-    end
+    @books = policy_scope(Book)
+    @other_books = Book.publishable - @books
   end
 
   # GET /books/1
   # GET /books/1.json
   def show
+  end
+
+  def view
+    @book = Book.find(params[:id])
+    @comments = @book.comments.recent.limit(10).all
+  end
+
+  def download
+    @book = Book.find(params[:id])
+    @chapters = @book.chapters.by_print_order
+    
+    respond_to do |format|
+      format.html
+      format.pdf do
+        render :pdf => "#{@book.title}",
+               
+               :layout                         => 'pdf.html.haml',
+               #:cover                          => "",
+               :page_size                      => "A5",
+               :font_name                      => "courier",
+               :header                         => { :font_size => 8, :center => @book.title},
+               :footer                         => { :center => '[page] of [topage]', :font_size => 8},
+               :disposition                    => 'attachment'
+      end
+    end
   end
 
   # GET /books/new
@@ -35,10 +56,64 @@ class BooksController < ApplicationController
 
     respond_to do |format|
       if @book.save
-        format.html { redirect_to books_url, notice: 'Book was successfully created.' }
-        format.json { render action: 'index', status: :created, location: books_url }
+        format.html { redirect_to book_url(@book), notice: 'Book was successfully created.' }
+        format.json { render action: 'index', status: :created, location: book_url(@book) }
       else
         format.html { render action: 'new' }
+        format.json { render json: @book.errors, status: :unprocessable_entity }
+      end
+    end
+  end
+
+  # PATCH/PUT /chapter/1
+  # PATCH/PUT /chapter/1.json
+  def publish
+    respond_to do |format|
+      if @book.publish!
+        format.html { redirect_to edit_book_path(@book), notice: 'Book successfully updated.' }
+        format.json { head :no_content }
+      else
+        format.html { redirect_to edit_book_path(@book), alert: "Book is not publishable. Try to set all chapters as completed and then try again" }
+        format.json { render json: @chapter.errors, status: :unprocessable_entity }
+      end
+    end
+  end
+
+  def unpublish
+    respond_to do |format|
+      if @book.unpublish!
+        format.html { redirect_to edit_book_path(@book), notice: 'Book successfully updated.' }
+        format.json { head :no_content }
+      else
+        format.html { redirect_to edit_book_path(@book), alert: "Book is not unpublishable." }
+        format.json { render json: @chapter.errors, status: :unprocessable_entity }
+      end
+    end
+  end
+
+  # PATCH/PUT /chapter/1
+  # PATCH/PUT /chapter/1.json
+  def complete
+    respond_to do |format|
+      if @book.complete!
+        format.html { redirect_to edit_book_url(@book), notice: 'Book was successfully updated.' }
+        format.json { head :no_content }
+      else
+        format.html { redirect_to edit_book_url(@book), alert: "Book was not completable. Check for the presence of #{LINK_REMOVED_TEXT} in the text" }
+        format.json { render json: @book.errors, status: :unprocessable_entity }
+      end
+    end
+  end
+
+  # PATCH/PUT /chapter/1
+  # PATCH/PUT /chapter/1.json
+  def uncomplete
+    respond_to do |format|
+      if @book.uncomplete!
+        format.html { redirect_to edit_book_url(@book), notice: 'Book was successfully updated.' }
+        format.json { head :no_content }
+      else
+        format.html { render action: 'edit' }
         format.json { render json: @book.errors, status: :unprocessable_entity }
       end
     end
@@ -49,7 +124,7 @@ class BooksController < ApplicationController
   def update
     respond_to do |format|
       if @book.update(book_params)
-        format.html { redirect_to books_url, notice: 'Book was successfully updated.' }
+        format.html { redirect_to book_url(@book), notice: 'Book was successfully updated.' }
         format.json { head :no_content }
       else
         format.html { render action: 'edit' }
@@ -72,10 +147,11 @@ class BooksController < ApplicationController
     # Use callbacks to share common setup or constraints between actions.
     def set_book
       @book = Book.find(params[:id])
+      authorize @book
     end
 
     # Never trust parameters from the scary internet, only allow the white list through.
     def book_params
-      params.require(:book).permit(:title, :cover,:user_id, :active)
+      params.require(:book).permit(:title, :introtext, :page_format, :font_name, :font_google, :font_google_name, :credits,  :cover,:user_id)
     end
 end
