@@ -1,64 +1,44 @@
 require 'spec_helper'
 
-# This spec was generated by rspec-rails when you ran the scaffold generator.
-# It demonstrates how one might use RSpec to specify the controller code that
-# was generated by Rails when you ran the scaffold generator.
-#
-# It assumes that the implementation code is generated by the rails scaffold
-# generator.  If you are using any extension libraries to generate different
-# controller code, this generated spec may or may not pass.
-#
-# It only uses APIs available in rails and/or rspec-rails.  There are a number
-# of tools you can use to make these specs even more expressive, but we're
-# sticking to rails and rspec-rails APIs to keep things simple and stable.
-#
-# Compared to earlier versions of this generator, there is very limited use of
-# stubs and message expectations in this spec.  Stubs are only used when there
-# is no simpler way to get a handle on the object needed for the example.
-# Message expectations are only used when there is no simpler way to specify
-# that an instance is receiving a specific message.
-
 describe BooksController do
   before :each do
-    login_user
+    login_user :user_with_book
   end
-  # This should return the minimal set of attributes required to create a valid
-  # Book. As you add validations to Book, be sure to
-  # adjust the attributes here as well.
-  let(:valid_attributes) { { "title" => "MyString" } }
 
-  # This should return the minimal set of values that should be in the session
-  # in order to pass any filters (e.g. authentication) defined in
-  # BooksController. Be sure to keep this updated too.
-  let(:valid_session) { {} }
+  let(:book) { @user.books.first}
+
+  before :each do
+    book.stub(:users).and_return([@user])
+  end
 
   describe "GET index" do
+    before :each do
+      @second_book = create(:book, :publishable)
+    end
     it "assigns all books as @books" do
-      book = Book.create! valid_attributes
-      get :index, {}, valid_session
+      get :index
       assigns(:books).should eq([book])
+      assigns(:other_books).should eq([@second_book])
     end
   end
 
   describe "GET show" do
     it "assigns the requested book as @book" do
-      book = Book.create! valid_attributes
-      get :show, {:id => book.to_param}, valid_session
+      get :show, {:id => book.to_param}
       assigns(:book).should eq(book)
     end
   end
 
   describe "GET new" do
     it "assigns a new book as @book" do
-      get :new, {}, valid_session
+      get :new
       assigns(:book).should be_a_new(Book)
     end
   end
 
   describe "GET edit" do
     it "assigns the requested book as @book" do
-      book = Book.create! valid_attributes
-      get :edit, {:id => book.to_param}, valid_session
+      get :edit, {:id => book.to_param}
       assigns(:book).should eq(book)
     end
   end
@@ -67,18 +47,18 @@ describe BooksController do
     describe "with valid params" do
       it "creates a new Book" do
         expect {
-          post :create, {:book => valid_attributes}, valid_session
+          post :create, {:book => book.attributes}
         }.to change(Book, :count).by(1)
       end
 
       it "assigns a newly created book as @book" do
-        post :create, {:book => valid_attributes}, valid_session
+        post :create, {:book => book.attributes}
         assigns(:book).should be_a(Book)
         assigns(:book).should be_persisted
       end
 
       it "redirects to the created book" do
-        post :create, {:book => valid_attributes}, valid_session
+        post :create, {:book => book.attributes}
         response.should redirect_to(Book.last)
       end
     end
@@ -87,14 +67,14 @@ describe BooksController do
       it "assigns a newly created but unsaved book as @book" do
         # Trigger the behavior that occurs when invalid params are submitted
         Book.any_instance.stub(:save).and_return(false)
-        post :create, {:book => { "title" => "invalid value" }}, valid_session
+        post :create, {:book => { "title" => "invalid value" }}
         assigns(:book).should be_a_new(Book)
       end
 
       it "re-renders the 'new' template" do
         # Trigger the behavior that occurs when invalid params are submitted
         Book.any_instance.stub(:save).and_return(false)
-        post :create, {:book => { "title" => "invalid value" }}, valid_session
+        post :create, {:book => { "title" => "invalid value" }}
         response.should render_template("new")
       end
     end
@@ -103,42 +83,37 @@ describe BooksController do
   describe "PUT update" do
     describe "with valid params" do
       it "updates the requested book" do
-        book = Book.create! valid_attributes
         # Assuming there are no other books in the database, this
         # specifies that the Book created on the previous line
         # receives the :update_attributes message with whatever params are
         # submitted in the request.
         Book.any_instance.should_receive(:update).with({ "title" => "MyString" })
-        put :update, {:id => book.to_param, :book => { "title" => "MyString" }}, valid_session
+        put :update, {:id => book.to_param, :book => { "title" => "MyString" }}
       end
 
       it "assigns the requested book as @book" do
-        book = Book.create! valid_attributes
-        put :update, {:id => book.to_param, :book => valid_attributes}, valid_session
+        put :update, {:id => book.to_param, :book => book.attributes}
         assigns(:book).should eq(book)
       end
 
       it "redirects to the book" do
-        book = Book.create! valid_attributes
-        put :update, {:id => book.to_param, :book => valid_attributes}, valid_session
+        put :update, {:id => book.to_param, :book => book.attributes}
         response.should redirect_to(book)
       end
     end
 
     describe "with invalid params" do
       it "assigns the book as @book" do
-        book = Book.create! valid_attributes
         # Trigger the behavior that occurs when invalid params are submitted
         Book.any_instance.stub(:save).and_return(false)
-        put :update, {:id => book.to_param, :book => { "title" => "invalid value" }}, valid_session
+        put :update, {:id => book.to_param, :book => { "title" => "invalid value" }}
         assigns(:book).should eq(book)
       end
 
       it "re-renders the 'edit' template" do
-        book = Book.create! valid_attributes
         # Trigger the behavior that occurs when invalid params are submitted
         Book.any_instance.stub(:save).and_return(false)
-        put :update, {:id => book.to_param, :book => { "title" => "invalid value" }}, valid_session
+        put :update, {:id => book.to_param, :book => { "title" => "invalid value" }}
         response.should render_template("edit")
       end
     end
@@ -146,17 +121,14 @@ describe BooksController do
 
   describe "DELETE destroy" do
     it "destroys the requested book" do
-      book = Book.create! valid_attributes
       expect {
-        delete :destroy, {:id => book.to_param}, valid_session
+        delete :destroy, {:id => book.to_param}
       }.to change(Book, :count).by(-1)
     end
 
     it "redirects to the books list" do
-      book = Book.create! valid_attributes
-      delete :destroy, {:id => book.to_param}, valid_session
+      delete :destroy, {:id => book.to_param}
       response.should redirect_to(books_url)
     end
   end
-
 end
