@@ -1,62 +1,40 @@
 require 'spec_helper'
 
-# This spec was generated by rspec-rails when you ran the scaffold generator.
-# It demonstrates how one might use RSpec to specify the controller code that
-# was generated by Rails when you ran the scaffold generator.
-#
-# It assumes that the implementation code is generated by the rails scaffold
-# generator.  If you are using any extension libraries to generate different
-# controller code, this generated spec may or may not pass.
-#
-# It only uses APIs available in rails and/or rspec-rails.  There are a number
-# of tools you can use to make these specs even more expressive, but we're
-# sticking to rails and rspec-rails APIs to keep things simple and stable.
-#
-# Compared to earlier versions of this generator, there is very limited use of
-# stubs and message expectations in this spec.  Stubs are only used when there
-# is no simpler way to get a handle on the object needed for the example.
-# Message expectations are only used when there is no simpler way to specify
-# that an instance is receiving a specific message.
-
 describe ChaptersController do
+  before :each do
+    login_user :user_with_book_and_chapters
+  end
 
-  # This should return the minimal set of attributes required to create a valid
-  # Chapter. As you add validations to Chapter, be sure to
-  # adjust the attributes here as well.
-  let(:valid_attributes) { { "title" => "MyString" } }
+  let(:chapter) { @user.books.first.chapters.first}
 
-  # This should return the minimal set of values that should be in the session
-  # in order to pass any filters (e.g. authentication) defined in
-  # ChaptersController. Be sure to keep this updated too.
-  let(:valid_session) { {} }
+  before :each do
+    
+  end
 
   describe "GET index" do
     it "assigns all chapters as @chapters" do
-      chapter = Chapter.create! valid_attributes
-      get :index, {}, valid_session
+      get :index, {:book_id => chapter.book.id}
       assigns(:chapters).should eq([chapter])
     end
   end
 
   describe "GET show" do
     it "assigns the requested chapter as @chapter" do
-      chapter = Chapter.create! valid_attributes
-      get :show, {:id => chapter.to_param}, valid_session
+      get :show, {:id => chapter.to_param, :book_id => chapter.book.id}
       assigns(:chapter).should eq(chapter)
     end
   end
 
   describe "GET new" do
     it "assigns a new chapter as @chapter" do
-      get :new, {}, valid_session
+      get :new, {:book_id => chapter.book.id}
       assigns(:chapter).should be_a_new(Chapter)
     end
   end
 
   describe "GET edit" do
     it "assigns the requested chapter as @chapter" do
-      chapter = Chapter.create! valid_attributes
-      get :edit, {:id => chapter.to_param}, valid_session
+      get :edit, {:id => chapter.to_param, :book_id => chapter.book.id}
       assigns(:chapter).should eq(chapter)
     end
   end
@@ -65,19 +43,19 @@ describe ChaptersController do
     describe "with valid params" do
       it "creates a new Chapter" do
         expect {
-          post :create, {:chapter => valid_attributes}, valid_session
+          post :create, {:book_id => chapter.book.id, :chapter => chapter.attributes}
         }.to change(Chapter, :count).by(1)
       end
 
       it "assigns a newly created chapter as @chapter" do
-        post :create, {:chapter => valid_attributes}, valid_session
+        post :create, {:book_id => chapter.book.id, :chapter => chapter.attributes}
         assigns(:chapter).should be_a(Chapter)
         assigns(:chapter).should be_persisted
       end
 
       it "redirects to the created chapter" do
-        post :create, {:chapter => valid_attributes}, valid_session
-        response.should redirect_to(Chapter.last)
+        post :create, {:book_id => chapter.book.id, :chapter => chapter.attributes}
+        response.should redirect_to(book_chapter_path(chapter.book, Chapter.last))
       end
     end
 
@@ -85,14 +63,14 @@ describe ChaptersController do
       it "assigns a newly created but unsaved chapter as @chapter" do
         # Trigger the behavior that occurs when invalid params are submitted
         Chapter.any_instance.stub(:save).and_return(false)
-        post :create, {:chapter => { "title" => "invalid value" }}, valid_session
+        post :create, {:book_id => chapter.book.id, :chapter => { "title" => "invalid value", "fulltext" => "" }}
         assigns(:chapter).should be_a_new(Chapter)
       end
 
       it "re-renders the 'new' template" do
         # Trigger the behavior that occurs when invalid params are submitted
         Chapter.any_instance.stub(:save).and_return(false)
-        post :create, {:chapter => { "title" => "invalid value" }}, valid_session
+        post :create, {:book_id => chapter.book.id, :chapter => { "title" => "invalid value", "fulltext" => "" }}
         response.should render_template("new")
       end
     end
@@ -101,42 +79,42 @@ describe ChaptersController do
   describe "PUT update" do
     describe "with valid params" do
       it "updates the requested chapter" do
-        chapter = Chapter.create! valid_attributes
+
         # Assuming there are no other chapters in the database, this
         # specifies that the Chapter created on the previous line
         # receives the :update_attributes message with whatever params are
         # submitted in the request.
         Chapter.any_instance.should_receive(:update).with({ "title" => "MyString" })
-        put :update, {:id => chapter.to_param, :chapter => { "title" => "MyString" }}, valid_session
+        put :update, {:book_id => chapter.book.id, :id => chapter.to_param, :chapter => { "title" => "MyString" }}
       end
 
       it "assigns the requested chapter as @chapter" do
-        chapter = Chapter.create! valid_attributes
-        put :update, {:id => chapter.to_param, :chapter => valid_attributes}, valid_session
+
+        put :update, {:book_id => chapter.book.id, :id => chapter.to_param, :chapter => chapter.attributes}
         assigns(:chapter).should eq(chapter)
       end
 
       it "redirects to the chapter" do
-        chapter = Chapter.create! valid_attributes
-        put :update, {:id => chapter.to_param, :chapter => valid_attributes}, valid_session
-        response.should redirect_to(chapter)
+
+        put :update, {:book_id => chapter.book.id, :id => chapter.to_param, :chapter => chapter.attributes}
+        response.should redirect_to(book_chapter_path(chapter.book, chapter))
       end
     end
 
     describe "with invalid params" do
       it "assigns the chapter as @chapter" do
-        chapter = Chapter.create! valid_attributes
+
         # Trigger the behavior that occurs when invalid params are submitted
         Chapter.any_instance.stub(:save).and_return(false)
-        put :update, {:id => chapter.to_param, :chapter => { "title" => "invalid value" }}, valid_session
+        put :update, {:book_id => chapter.book.id, :id => chapter.to_param, :chapter => { "title" => "invalid value" }}
         assigns(:chapter).should eq(chapter)
       end
 
       it "re-renders the 'edit' template" do
-        chapter = Chapter.create! valid_attributes
+
         # Trigger the behavior that occurs when invalid params are submitted
         Chapter.any_instance.stub(:save).and_return(false)
-        put :update, {:id => chapter.to_param, :chapter => { "title" => "invalid value" }}, valid_session
+        put :update, {:book_id => chapter.book.id, :id => chapter.to_param, :chapter => { "title" => "invalid value" }}
         response.should render_template("edit")
       end
     end
@@ -144,16 +122,14 @@ describe ChaptersController do
 
   describe "DELETE destroy" do
     it "destroys the requested chapter" do
-      chapter = Chapter.create! valid_attributes
       expect {
-        delete :destroy, {:id => chapter.to_param}, valid_session
+        delete :destroy, {:book_id => chapter.book.id, :id => chapter.to_param}
       }.to change(Chapter, :count).by(-1)
     end
 
     it "redirects to the chapters list" do
-      chapter = Chapter.create! valid_attributes
-      delete :destroy, {:id => chapter.to_param}, valid_session
-      response.should redirect_to(chapters_url)
+      delete :destroy, {:book_id => chapter.book.id, :id => chapter.to_param}
+      redirect_to(book_chapters_path(chapter.book))
     end
   end
 
