@@ -2,13 +2,30 @@
 
 FactoryGirl.define do
   factory :chapter do
-    title "MyString"
-    introtext "MyText"
-    fulltext "MyText"
-    active false
+    sequence(:id)
+    title "Chapter"
+    introtext "IntroText"
+    fulltext "FullText"
+    completed false
     death false
     ending false
     beginning false
     book nil
+  end
+
+  factory :chapter_completed, :parent => :chapter do
+    completed true
+  end
+
+  factory :chapter_uncompleted, :parent => :chapter do
+    completed false
+  end
+
+  factory :chapter_with_book, :parent => :chapter do
+    book { |b| b.association(:book) }
+  end
+
+  factory :chapter_with_user, :parent => :chapter do
+    book { |b| b.association(:book_with_user) }
   end
 end
