@@ -5,7 +5,7 @@ describe "books/edit" do
     @book = assign(:book, stub_model(Book,
       :title => "MyString",
       :user_id => 1,
-      :active => ""
+      :completed => ""
     ))
   end
 
@@ -15,7 +15,6 @@ describe "books/edit" do
     # Run the generator again with the --webrat flag if you want to use webrat matchers
     assert_select "form[action=?][method=?]", book_path(@book), "post" do
       assert_select "input#book_title[name=?]", "book[title]"
-      assert_select "input[name=?]", "book[active]"
     end
   end
 end
